+++ flutter_inappwebview_forge_platform_interface/lib/src/types/main.dart
@@ -186,6 +186,7 @@ export 'url_request_network_service_type.dart'
     show URLRequestNetworkServiceType, IOSURLRequestNetworkServiceType;
 export 'url_response.dart' show URLResponse, IOSURLResponse;
 export 'user_preferred_content_mode.dart' show UserPreferredContentMode;
+export 'writing_tools_behavior.dart' show IOSWritingToolsBehavior;
 export 'user_script.dart' show UserScript;
 export 'user_script_injection_time.dart' show UserScriptInjectionTime;
 export 'vertical_scrollbar_position.dart'